+++ weblate/machinery/weblatetm.py
@@ -31,6 +31,7 @@ class WeblateTranslation(MachineTranslation):
     name = "Weblate"
     rank_boost = 1
     cache_translations = False
+    do_cleanup = False
 
     def convert_language(self, language):
         """No conversion of language object."""