+++ weblate/memory/machine.py
@@ -29,6 +29,7 @@ class WeblateMemory(MachineTranslation):
     rank_boost = 2
     cache_translations = False
     same_languages = True
+    do_cleanup = False
 
     def convert_language(self, language):
         """No conversion of language object."""